+++ NEWS.md
@@ -2,6 +2,8 @@
 ## Added
 * Added new internal function `SplitOneVCF`, `SplitListOfVCFs`.
 
+* Added new *exported* function `ReadAndSplitVCFs`.
+
 ## Changed
 * Updated function `MakeDataFrameFromVCF` to use `data.table::fread` instead of 
 `read.csv`.